guard against strings in #label_with_preferred_language
As of https://github.com/ActiveTriples/ActiveTriples/commit/e6d6f69,
the `{ literal: true }` option is no longer recognized, and strings
are not automatically converted to `RDF::Literals`.

For e.g., applications manually indexing the labels on their local
authorities, that process now fails since `get_values` returns a an
array of strings which do not respond to `#language`.